pyproject.toml
[tool.poetry]
name = "automate"
version = "0.1.0"
description = ""
authors = ["yuruotong"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
PyYAML = "6.0.1"
requests = "2.31.0"
selenium = "4.16.0"
webdriver-manager = "4.0.1"
pytest = "^8.0.1"
pyqt6 = "^6.6.1"
pyautogui = "^0.9.54"
pyinstaller = "^6.5.0"
pygments = "^2.17.2"
python-docx = "^1.1.2"
pynput = "^1.7.7"
open-interpreter = "^0.2.5"
openai = "^1.30.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "default"
url = 'https://pypi.org/'
priority = "primary"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "supplemental"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "supplemental"